my-family-app/src/core/services/auth/passwordService.ts
// src/core/services/auth/passwordService.ts
export interface PasswordResetResponse {
    success: boolean;
    message?: string;
  }
  
  export interface IPasswordService {
    requestPasswordReset(email: string): Promise<PasswordResetResponse>;
    resetPassword(token: string, newPassword: string): Promise<PasswordResetResponse>;
  }
  
  export const PasswordService: IPasswordService = {
    async requestPasswordReset(email: string): Promise<PasswordResetResponse> {
      // TODO: Implement actual API call
      return new Promise((resolve) => {
        setTimeout(() => {
          console.log(`Password reset requested for ${email}`);
          resolve({
            success: true,
            message: 'Password reset instructions sent to your email'
          });
        }, 1000);
      });
    },
  
    async resetPassword(token: string, password: string): Promise<PasswordResetResponse> {
      // TODO: Implement actual API call
      return new Promise((resolve) => {
        setTimeout(() => {
          console.log(`Password reset with token ${token} and new password length: ${password.length}`);
          resolve({
            success: true,
            message: 'Password has been reset successfully'
          });
        }, 1000);
      });
    }
  };